Syntax: Create syntax_dev package
