+++ plugins/syntax_dev/__init__.py
@@ -1,0 +1,2 @@
+from .completions import *
+from .highlighter import *
