+++ plugins/syntax_dev/completions.py
@@ -4,16 +4,25 @@ import re
 import sublime
 import sublime_plugin
 
-from sublime_lib.flags import RegionOption
-
-from .lib.scope_data import COMPILED_HEADS
-from .lib import syntax_paths
+from ..lib.scope_data import COMPILED_HEADS
+from ..lib import syntax_paths
 
 __all__ = (
-    'SyntaxDefRegexCaptureGroupHighlighter',
     'SyntaxDefCompletionsListener',
     'PackagedevCommitScopeCompletionCommand'
 )
+
+
+# a list of kinds used to denote the different kinds of completions
+KIND_HEADER_BASE = (sublime.KIND_ID_NAMESPACE, 'K', 'Header Key')
+KIND_HEADER_DICT = (sublime.KIND_ID_NAMESPACE, 'D', 'Header Dict')
+KIND_HEADER_LIST = (sublime.KIND_ID_NAMESPACE, 'L', 'Header List')
+KIND_BRANCH = (sublime.KIND_ID_NAVIGATION, 'b', 'Branch Point')
+KIND_CONTEXT = (sublime.KIND_ID_KEYWORD, 'c', 'Context')
+KIND_FUNCTION = (sublime.KIND_ID_FUNCTION, 'f', 'Function')
+KIND_CAPTURUE = (sublime.KIND_ID_FUNCTION, 'c', 'Captures')
+KIND_SCOPE = (sublime.KIND_ID_NAMESPACE, 's', 'Scope')
+KIND_VARIABLE = (sublime.KIND_ID_VARIABLE, 'v', 'Variable')
 
 PACKAGE_NAME = __package__.split('.')[0]
 
@@ -23,98 +32,6 @@ def status(msg, console=False):
     sublime.status_message(msg)
     if console:
         print(msg)
-
-
-class SyntaxDefRegexCaptureGroupHighlighter(sublime_plugin.ViewEventListener):
-
-    # TODO multiple views into the same file
-
-    @classmethod
-    def is_applicable(cls, settings):
-        return settings.get('syntax') == syntax_paths.SYNTAX_DEF
-
-    def on_selection_modified(self):
-        prefs = sublime.load_settings('PackageDev.sublime-settings')
-        scope = prefs.get('syntax.captures_highlight_scope', "text")
-        styles = prefs.get('syntax.captures_highlight_styles', ['DRAW_NO_FILL'])
-
-        style_flags = RegionOption(*styles)
-
-        self.view.add_regions(
-            key='captures',
-            regions=list(self.get_regex_regions()),
-            scope=scope,
-            flags=style_flags,
-        )
-
-    def get_regex_regions(self):
-        locations = [
-            region.begin()
-            for selection in self.view.sel()
-            if self.view.match_selector(
-                selection.begin(),
-                'source.yaml.sublime.syntax meta.expect-captures'
-            )
-            for region in self.view.split_by_newlines(selection)
-        ]
-
-        for loc in locations:
-            # Find the line number.
-            match = re.search(r'(\d+):', self.view.substr(self.view.line(loc)))
-            if not match:
-                continue
-            n = int(match.group(1))
-
-            # Find the associated regexp. Assume it's the preceding one.
-            try:
-                regexp_region = [
-                    region
-                    for region in self.view.find_by_selector('source.regexp.oniguruma')
-                    if region.end() < loc
-                ][-1]
-            except IndexError:
-                continue
-
-            if n == 0:
-                yield regexp_region
-                continue
-
-            # Find parens that define capture groups.
-            regexp_offset = regexp_region.begin()
-            parens = iter(
-                (match.group(), match.start() + regexp_offset)
-                for match in re.finditer(r'\(\??|\)', self.view.substr(regexp_region))
-                if self.view.match_selector(
-                    match.start() + regexp_offset,
-                    'keyword.control.group'
-                )
-            )
-
-            # Find the start of the nth capture group.
-            start = None
-            count = 0
-            for p, i in parens:
-                if p == '(':  # Not (?
-                    count += 1
-                    if count == n:
-                        start = i
-                        break
-
-            # Find the end of that capture group
-            end   = None
-            depth = 0
-            for p, i in parens:
-                if p in {'(', '(?'}:
-                    depth += 1
-                else:
-                    if depth == 0:
-                        end = i + 1
-                        break
-                    else:
-                        depth -= 1
-
-            if end is not None:
-                yield sublime.Region(start, end)
 
 
 def _inhibit_word_completions(func):
@@ -127,107 +44,56 @@ def _inhibit_word_completions(func):
     return wrapper
 
 
-HAVE_KINDS = hasattr(sublime, 'CompletionItem')
-if HAVE_KINDS:
+def format_static_completions(templates):
 
-    # a list of kinds used to denote the different kinds of completions
-    KIND_HEADER_BASE = (sublime.KIND_ID_NAMESPACE, 'K', 'Header Key')
-    KIND_HEADER_DICT = (sublime.KIND_ID_NAMESPACE, 'D', 'Header Dict')
-    KIND_HEADER_LIST = (sublime.KIND_ID_NAMESPACE, 'L', 'Header List')
-    KIND_BRANCH = (sublime.KIND_ID_NAVIGATION, 'b', 'Branch Point')
-    KIND_CONTEXT = (sublime.KIND_ID_KEYWORD, 'c', 'Context')
-    KIND_FUNCTION = (sublime.KIND_ID_FUNCTION, 'f', 'Function')
-    KIND_CAPTURUE = (sublime.KIND_ID_FUNCTION, 'c', 'Captures')
-    KIND_SCOPE = (sublime.KIND_ID_NAMESPACE, 's', 'Scope')
-    KIND_VARIABLE = (sublime.KIND_ID_VARIABLE, 'v', 'Variable')
+    def format_item(trigger, kind, details):
+        if kind in (KIND_HEADER_DICT, KIND_CAPTURUE, KIND_CONTEXT):
+            completion_format = sublime.COMPLETION_FORMAT_SNIPPET
+            suffix = ":\n  "
+        elif kind is KIND_HEADER_LIST:
+            completion_format = sublime.COMPLETION_FORMAT_SNIPPET
+            suffix = ":\n  - "
+        else:
+            completion_format = sublime.COMPLETION_FORMAT_TEXT
+            suffix = ": "
 
-    def format_static_completions(templates):
+        return sublime.CompletionItem(
+            trigger=trigger, kind=kind, details=details,
+            completion=trigger + suffix, completion_format=completion_format
+        )
 
-        def format_item(trigger, kind, details):
-            if kind in (KIND_HEADER_DICT, KIND_CAPTURUE, KIND_CONTEXT):
-                completion_format = sublime.COMPLETION_FORMAT_SNIPPET
-                suffix = ":\n  "
-            elif kind is KIND_HEADER_LIST:
-                completion_format = sublime.COMPLETION_FORMAT_SNIPPET
-                suffix = ":\n  - "
-            else:
-                completion_format = sublime.COMPLETION_FORMAT_TEXT
-                suffix = ": "
+    return [format_item(*template) for template in templates]
 
-            return sublime.CompletionItem(
-                trigger=trigger, kind=kind, details=details,
-                completion=trigger + suffix, completion_format=completion_format
-            )
 
-        return [format_item(*template) for template in templates]
+def format_completions(items, annotation="", kind=sublime.KIND_AMBIGUOUS):
+    format_string = "Defined at line <a href='subl:goto_line {{\"line\": \"{0}\"}}'>{0}</a>"
+    return [
+        sublime.CompletionItem(
+            trigger=trigger, annotation=annotation, kind=kind,
+            details=format_string.format(row) if row is not None else ""
+        )
+        for trigger, row in items
+    ]
 
-    def format_completions(items, annotation="", kind=sublime.KIND_AMBIGUOUS):
-        format_string = "Defined at line <a href='subl:goto_line {{\"line\": \"{0}\"}}'>{0}</a>"
-        return [
-            sublime.CompletionItem(
-                trigger=trigger, annotation=annotation, kind=kind,
-                details=format_string.format(row) if row is not None else ""
-            )
-            for trigger, row in items
-        ]
 
-    def format_branch_completions(items):
-        return format_completions(items, "", KIND_BRANCH)
+def format_branch_completions(items):
+    return format_completions(items, "", KIND_BRANCH)
 
-    def format_context_completions(items):
-        return format_completions(items, "", KIND_CONTEXT)
 
-    def format_base_completion(item):
-        return format_completions([[item, None], ], "base suffix", KIND_SCOPE)
+def format_context_completions(items):
+    return format_completions(items, "", KIND_CONTEXT)
 
-    def format_scope_completions(items):
-        return format_completions(items, "convention", KIND_SCOPE)
 
-    def format_variable_completions(items):
-        return format_completions(items, "", KIND_VARIABLE)
+def format_base_completion(item):
+    return format_completions([[item, None], ], "base suffix", KIND_SCOPE)
 
-else:
 
-    KIND_HEADER_BASE = ''
-    KIND_HEADER_DICT = ''
-    KIND_HEADER_LIST = ''
-    KIND_BRANCH = ''
-    KIND_CONTEXT = ''
-    KIND_FUNCTION = ''
-    KIND_CAPTURUE = ''
-    KIND_VARIABLE = ''
+def format_scope_completions(items):
+    return format_completions(items, "convention", KIND_SCOPE)
 
-    def format_static_completions(templates):
 
-        def format_item(trigger, kind, details):
-            if kind in (KIND_HEADER_DICT, KIND_CAPTURUE, KIND_CONTEXT):
-                suffix = ":\n  "
-            elif kind is KIND_HEADER_LIST:
-                suffix = ":\n  - "
-            else:
-                suffix = ": "
-
-            return ("{0}\t{1}".format(trigger, details), trigger + suffix)
-
-        return [format_item(*template) for template in templates]
-
-    def format_completions(items, kind):
-        return [trigger + kind for trigger, _ in items]
-
-    def format_base_completion(item):
-        return [item + "\tbase suffix"]
-
-    def format_branch_completions(items):
-        return format_completions(items, "\tbranch point")
-
-    def format_context_completions(items):
-        return format_completions(items, "\tcontext")
-
-    def format_scope_completions(items):
-        return format_completions(items, "\tconvention")
-
-    def format_variable_completions(items):
-        return format_completions(items, "\tvariable")
+def format_variable_completions(items):
+    return format_completions(items, "", KIND_VARIABLE)
 
 
 class SyntaxDefCompletionsListener(sublime_plugin.ViewEventListener):
